refactoring / adjust container, presenter design pattern
